[Agent] move resolution source builder
